3 Changes
* Change Virtual Project sort priority so it shows between "Solution Items" and regular projects.
* Use custom icons for root folder, project folder and files.
* Implement "Open With..." context menu item.

+++ src/VsChromium/Features/SourceExplorerHierarchy/NodeViewModel.cs
@@ -220,14 +220,14 @@ namespace VsChromium.Features.SourceExplorerHierarchy {
               iconHandle = GetIconHandle();
             if (iconHandle == IntPtr.Zero)
               return VSConstants.E_NOTIMPL;
-            pvar = iconHandle;
+            pvar = (int)iconHandle;
             break;
           }
         case (int)__VSHPROPID.VSHPROPID_IconHandle: {
             var iconHandle = GetIconHandle();
             if (iconHandle == IntPtr.Zero)
               return VSConstants.E_NOTIMPL;
-            pvar = iconHandle;
+            pvar = (int)iconHandle;
             break;
           }
         case (int)__VSHPROPID.VSHPROPID_ProjectName:
